+++ tests/test_central.py
@@ -17,9 +17,14 @@
 #
 """Test functionalty of OdkCentral.py."""
 
+from io import BytesIO
 from pathlib import Path
 
+import pytest
+import requests
 import segno
+
+testdata_dir = Path(__file__).parent / "testdata"
 
 
 def test_delete_appuser(appuser, appuser_details, project_details):
@@ -70,3 +75,152 @@ def test_create_qrcode(appuser, appuser_details):
     )
     qrcode_file = Path("test project.png")
     assert qrcode_file.exists()
+
+
+def test_create_form_delete(project, odk_form):
+    """Create form and delete."""
+    odk_id, xform = odk_form
+    test_xform = testdata_dir / "buildings.xml"
+
+    form_name = xform.createForm(odk_id, str(test_xform))
+    assert form_name == "test_form"
+
+    assert len(project.listForms(odk_id)) == 1
+
+    success = xform.deleteForm(odk_id, form_name)
+    assert success
+
+    assert len(project.listForms(odk_id)) == 0
+
+
+def test_create_form_and_publish(project, odk_form_cleanup):
+    """Create form and publish."""
+    odk_id, form_name, xform = odk_form_cleanup
+
+    response_code = xform.publishForm(odk_id, form_name)
+    assert response_code == 200
+    assert xform.published == True
+
+
+def test_create_form_and_publish_immediately(project, odk_form):
+    """Create form and publish immediately."""
+    odk_id, xform = odk_form
+    test_xform = testdata_dir / "buildings.xml"
+
+    form_name = xform.createForm(odk_id, str(test_xform), publish=True)
+    assert form_name == "test_form"
+    assert xform.published == True
+
+    success = xform.deleteForm(odk_id, form_name)
+    assert success
+
+    assert len(project.listForms(odk_id)) == 0
+
+
+def test_create_form_draft(project, odk_form_cleanup):
+    """Create form draft from existing form."""
+    odk_id, original_form_name, xform = odk_form_cleanup
+    test_xform = testdata_dir / "buildings.xml"
+
+    # Check original form is not draft
+    assert xform.draft == False
+
+    # Publish original form
+    response_code = xform.publishForm(odk_id, original_form_name)
+    assert response_code == 200
+    assert xform.published == True
+
+    # Create draft from original form (sleep 1s first for version increment)
+    draft_form_name = xform.createForm(odk_id, str(test_xform), original_form_name)
+    assert draft_form_name == original_form_name
+    assert xform.draft == True
+
+    # Delete the newly created draft
+    success = xform.deleteForm(odk_id, draft_form_name)
+    assert success
+
+    # Create another draft from original form
+    draft_form_name = xform.createForm(odk_id, str(test_xform), original_form_name)
+    assert draft_form_name == original_form_name
+    assert xform.draft == True
+
+    # Publish newly created version of form
+    response_code = xform.publishForm(odk_id, draft_form_name)
+    assert response_code == 200
+    assert xform.published == True
+    assert xform.draft == False
+
+    assert len(project.listForms(odk_id)) == 1
+
+
+def test_upload_media_filepath(project, odk_form_cleanup):
+    """Create form and upload media."""
+    odk_id, form_name, xform = odk_form_cleanup
+
+    # Publish form first
+    response_code = xform.publishForm(odk_id, form_name)
+    assert response_code == 200
+    assert xform.published == True
+
+    # Upload media
+    result = xform.uploadMedia(
+        odk_id,
+        "test_form",
+        str(testdata_dir / "osm_buildings.geojson"),
+    )
+    assert result.status_code == 200
+
+
+def test_upload_media_bytesio_publish(project, odk_form):
+    """Create form and upload media."""
+    odk_id, xform = odk_form
+    test_xform = testdata_dir / "buildings.xml"
+    with open(test_xform, "rb") as xform_file:
+        xform_bytesio = BytesIO(xform_file.read())
+
+    # Create form
+    form_name = xform.createForm(odk_id, xform_bytesio)
+    assert form_name == "test_form"
+
+    # Publish form first
+    response_code = xform.publishForm(odk_id, form_name)
+    assert response_code == 200
+    assert xform.published == True
+
+    # Upload media
+    with open(testdata_dir / "osm_buildings.geojson", "rb") as geojson:
+        geojson_bytesio = BytesIO(geojson.read())
+    result = xform.uploadMedia(odk_id, "test_form", geojson_bytesio, filename="osm_buildings.geojson")
+    assert result.status_code == 200
+
+    # Delete form
+    success = xform.deleteForm(odk_id, "test_form")
+    assert success
+
+    assert len(project.listForms(odk_id)) == 0
+
+
+def test_form_fields_no_form(odk_form):
+    """Attempt usage of form_fields before form exists."""
+    odk_id, xform = odk_form
+    with pytest.raises(requests.exceptions.HTTPError):
+        xform.formFields(odk_id, "test_form")
+
+
+def test_form_fields(odk_form_cleanup):
+    """Test form fields for created form."""
+    odk_id, form_name, xform = odk_form_cleanup
+
+    # Get form fields
+    form_fields = xform.formFields(odk_id, form_name)
+    assert len(form_fields) == 66
+
+    sorted_form_fields = sorted(form_fields, key=lambda x: x["name"])
+    buildings_heritage = sorted_form_fields[30]
+    assert buildings_heritage == {
+        "path": "/all/buildings/heritage",
+        "name": "heritage",
+        "type": "string",
+        "binary": None,
+        "selectMultiple": None,
+    }